Reads the output of a function calculated on the Slurm cluster

This function reads all function output files (one by cluster node used) from the specified Slurm job and returns the result in a single data frame (if "table" format selected) or list (if "raw" format selected). It doesn't record any messages (including warnings or errors) output to the R console during the computation; these can be consulted by invoking print_job_status.

get_slurm_out(slr_job, outtype = "raw", wait = TRUE, ncores = NULL)

Arguments

slr_job

A slurm_job object.

outtype

Can be "table" or "raw", see "Value" below for details.

wait

Specify whether to block until slr_job completes.

ncores

(optional) If not null, the number of cores passed to mclapply

Value

If outtype = "table": A data frame with one column by return value of the function passed to slurm_apply, where each row is the output of the corresponding row in the params data frame passed to slurm_apply.

If outtype = "raw": A list where each element is the output of the function passed to slurm_apply for the corresponding row in the params data frame passed to slurm_apply.

Details

The outtype option is only relevant for jobs submitted with slurm_apply. Jobs sent with slurm_call only return a single object, and setting outtype = "table" creates an error in that case.

See also